issue-38509 修改商品数量字段类型

+++ db/cloudprint1.4-delta-schema.sql
@@ -46,7 +46,7 @@ ALTER TABLE `eh_payment_bill_items` ADD COLUMN `goods_serve_apply_name` VARCHAR(
 ALTER TABLE `eh_payment_bill_items` ADD COLUMN `goods_tag` VARCHAR(1024) COMMENT '商品标识，如：活动ID、商品ID';
 ALTER TABLE `eh_payment_bill_items` ADD COLUMN `goods_name` VARCHAR(1024) COMMENT '商品名称';
 ALTER TABLE `eh_payment_bill_items` ADD COLUMN `goods_description` VARCHAR(1024) COMMENT '商品说明';
-ALTER TABLE `eh_payment_bill_items` ADD COLUMN `goods_counts` BIGINT COMMENT '商品数量';
+ALTER TABLE `eh_payment_bill_items` ADD COLUMN `goods_counts` INTEGER COMMENT '商品数量';
 ALTER TABLE `eh_payment_bill_items` ADD COLUMN `goods_price` DECIMAL(10,2) COMMENT '商品单价';
 ALTER TABLE `eh_payment_bill_items` ADD COLUMN `goods_totalPrice` DECIMAL(10,2) COMMENT '商品总金额';
 
@@ -61,7 +61,7 @@ ALTER TABLE `eh_asset_module_app_mappings` ADD COLUMN `goods_serve_apply_name` V
 ALTER TABLE `eh_asset_module_app_mappings` ADD COLUMN `goods_tag` VARCHAR(1024) COMMENT '商品标识，如：活动ID、商品ID';
 ALTER TABLE `eh_asset_module_app_mappings` ADD COLUMN `goods_name` VARCHAR(1024) COMMENT '商品名称';
 ALTER TABLE `eh_asset_module_app_mappings` ADD COLUMN `goods_description` VARCHAR(1024) COMMENT '商品说明';
-ALTER TABLE `eh_asset_module_app_mappings` ADD COLUMN `goods_counts` BIGINT COMMENT '商品数量';
+ALTER TABLE `eh_asset_module_app_mappings` ADD COLUMN `goods_counts` INTEGER COMMENT '商品数量';
 ALTER TABLE `eh_asset_module_app_mappings` ADD COLUMN `goods_price` DECIMAL(10,2) COMMENT '商品单价';
 ALTER TABLE `eh_asset_module_app_mappings` ADD COLUMN `goods_totalPrice` DECIMAL(10,2) COMMENT '商品总金额';
